refactor: Update World class to use JSON serialization for WorldSettings
This commit updates the World class in the Deesix.Domain.Entities namespace to use JSON serialization for the WorldSettings property. Instead of directly storing the WorldSettings object, the World class now has a WorldSettingsJson property that stores the serialized JSON string. The WorldSettings property is now a computed property that deserializes the JSON string into a WorldSettings object when accessed.

- Add WorldSettingsJson property to World class
- Add computed WorldSettings property that uses JSON deserialization